preferring long

+++ src/main/java/com/marincic/assignment/model/BonusCard.java
@@ -2,6 +2,6 @@ package com.marincic.assignment.model;
 
 public record BonusCard(
         Integer id,
-        int numberOfBeveragesPurchased
+        long numberOfBeveragesPurchased
 ) {
 }
